prov/rxm: Move rxm_av.c to util code
The AV will be shared by the tcp provider for rdm ep support.
Move the source file and necessary definitions to the util
code.
