+++ libfabric.vcxproj
@@ -445,7 +445,7 @@
     <ClCompile Include="prov\rxm\src\rxm_tagged.c" />
     <ClCompile Include="prov\rxm\src\rxm_rma.c" />
     <ClCompile Include="prov\rxm\src\rxm_cq.c" />
-    <ClCompile Include="prov\rxm\src\rxm_av.c" />
+    <ClCompile Include="prov\util\src\rxm_av.c" />
     <ClCompile Include="prov\rxm\src\rxm_domain.c" />
     <ClCompile Include="prov\rxm\src\rxm_ep.c" />
     <ClCompile Include="prov\rxm\src\rxm_fabric.c" />
