+++ prov/rxm/Makefile.include
@@ -9,7 +9,6 @@ _rxm_files = \
        prov/rxm/src/rxm_msg.c	\
        prov/rxm/src/rxm_tagged.c	\
        prov/rxm/src/rxm_cq.c	\
-       prov/rxm/src/rxm_av.c	\
        prov/rxm/src/rxm_rma.c	\
        prov/rxm/src/rxm_atomic.c	\
        prov/rxm/src/rxm.h
